thêm hệ thống thông minh

+++ SpringBoot-CuaHangDienThoai/src/main/java/com/cuahangdienthoai/service/RecommendationService.java
@@ -1,7 +1,7 @@
 package com.cuahangdienthoai.service;
 
 public interface RecommendationService {
-    String getRecommendOfDevices(int productID);
+    String getRecommendOfDevices(Long productID);
     String getRecommendOfUser(int userId);
 
     void train();
